requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.98.0
frozenlist==1.3.3
googlesearch-python==1.2.3
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httptools==0.5.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
openai==0.27.8
passlib==1.7.4
people-also-ask==1.1.0
pydantic==1.10.9
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.7.0
tzlocal==5.0.1
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
